fix Creative Commons licenses' classification
CC licenses are not intended to be used for software, and therefore do
not contain any definition of "source code" nor any provision related to
"source code". This implies that:

- it is not correct to classify NC and ND licenses as
"source-available"; rather, they should be classified as
"free-restricted" (one or more of the 4 freedoms are not fully granted
or are limited)

- it is not correct to classify SA licenses as
"include-source-code-offer-in-notice-file" (there is not, and cannot be,
any such provision in the license)

Moreover, SA licenses cannot be classified as "copyleft-limited" since
they are "copyleft" licenses in all respects, with no exception granted;
the fact that a work other than software is involved may lead to
different practical implications when dealing with derivative works, but
that does not imply that there are any exceptions to the copyleft
clauses.

Signed-off-by: Alberto Pianon <alberto@pianon.eu>