Change JSON license from "permissive" to "free-restricted"
This is debatable, because the condition "The Software shall be used for
Good, not Evil" may be intended as an unenforceable soft requirement.
However, to err on the safe side, we interpret it as a restriction, in
order to allow the auditor to decide on a case-by-case basis whether it
poses an actual risk or not.

Signed-off-by: Alberto Pianon <alberto@pianon.eu>