Change JSON license from "permissive" to "free-restricted"
The issue is the condition "The Software shall be used for Good, not
Evil", which makes the JSON license fall under the category of "ethical
licenses", which are contrary to the OSD and the fundamental tenet of
Free Software. Actually, the use of "shall" instead of "should" makes it
a legal requirement of unpredictable consequences, even if some
interpret it as an unenforceable soft requirement. To err on the safe
side, we categorize this license as "free-restricted", in order to allow
the auditor to decide on a case-by-case basis whether it poses an actual
risk or not.

Reviewed-by: Carlo Piana <carlo@piana.eu>
Signed-off-by: Alberto Pianon <alberto@pianon.eu>