Add entry for LicenseRef-scancode-ecma-no-patent
Even if it is a notice and not a license, it is important to flag it as
"patent-license" because that category is intended to raise warnings
when acquiring a patent license may be required. It probably bears no
consequences in licensing, since it speaks of rights which are not
controlled by the licensor(s), but it is indicative of a potential
problem that, once seen, cannot be ignored.

Reviewed-by: Carlo Piana <carlo@piana.eu>
Signed-off-by: Alberto Pianon <alberto@pianon.eu>