Change unrar licenses from source-available to free-restricted
This may be debatable, but by and large "source-available" refers to a
specific phenomenon of software claiming to be Open Source but not
satisfying the OSD requirements (see
https://en.wikipedia.org/wiki/Source-available_software), or more generally to anything that is "not Open Source"; unrar never
claimed to be Open Source, and its license grants some
freedoms but with limitations and restrictions (so "free-restricted").
The risks normally associated with "source-available" and "free-restricted" licenses are different, and the former are generally
considered more dangerous than the latter, as they are closer to a
fully proprietary licensing scheme.

Reviewed-by: Carlo Piana <carlo@piana.eu>
Signed-off-by: Alberto Pianon <alberto@pianon.eu>