Add handler and rule for free-restricted-copyleft licenses
The category "free-restricted-copyleft" is mainly intended to handle
CC-(ND|NC)-SA-* licenses, with are both copyleft and free-restricted
licenses. Since the logic of the standard evaluator.rules.kts does not
allow to have multiple categories for the same license, which may
trigger multiple rules (it would raise a duplicate license error), we
created a dedicated category in license-classification.yml, and here
we create the necessary handler and rule.

Signed-off-by: Alberto Pianon <alberto@pianon.eu>